requirements.txt
flask
pandas
numpy
scikit-learn
vaderSentiment
python-dotenv
flask-limiter
